+++ src/main/kotlin/com/github/andrelmv/kotbox/actions/wrappers/WrapCoroutineScopeAction.kt
@@ -1,0 +1,5 @@
+package com.github.andrelmv.kotbox.actions.wrappers
+
+import com.github.andrelmv.kotbox.actions.WrapWithCoroutineAction
+
+class WrapCoroutineScopeAction : WrapWithCoroutineAction("coroutineScope")